Update the planet size definitions for WarpDrive
I updated the radiusX and radiusZ to the proper values, 2500 - the polygon radius for the inscribed square of the world border circle - so that space will be enterrable from more than just a tiny area around the gate.

Random09 also did this. However, he used the Width and Height instead of polygon radius. Both of us assumed we were right(I because it used radiusX and radiusZ, he because it would be simpler to use Width and Height), so we both submitted. We assume that you know which is actually correct, so implement that one. Either way, we will now be able to enter space without being infernally close to the stargate. A large portion of the space currently inside of the planets' definitions are in the protected zone.